feat: Added class validator and adjust contacts macapa

+++ src/contacts-macapa/entities/contacts-macapa.entity.ts
@@ -2,7 +2,7 @@ import { BaseEntity, Column, Entity, PrimaryGeneratedColumn } from 'typeorm';
 
 @Entity({ name: 'contacts' })
 export class ContactsMacapa extends BaseEntity {
-  @PrimaryGeneratedColumn({ type: 'bigint', unsigned: true })
+  @PrimaryGeneratedColumn({ type: 'int', unsigned: true })
   id: number;
 
   @Column({ length: 200, nullable: false })
